requirements.txt
beautifulsoup4==4.8.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Flask==1.1.1
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
lxml==4.4.1
MarkupSafe==1.1.1
oauthlib==3.1.0
PySocks==1.7.1
python-dotenv==0.10.3
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
soupsieve==1.9.4
tweepy==3.8.0
tzlocal==2.0.0
urllib3==1.25.6
Werkzeug==0.16.0
